wifi: rtw88: prepare core for RTL8188E descriptors
RTL8188E uses the older 32-byte TX and 24-byte RX descriptor formats
rather than the formats currently handled by the rtw88 core. Add optional
chip callbacks for both directions and support its in-band TX status
reports.

The I-cut power sequence also needs a cut bit beyond the existing eight-bit
mask. Widen the mask and describe H- and I-cut explicitly.

+++ drivers/net/wireless/realtek/rtw88/mac.c
@@ -183,7 +183,7 @@ err:
 }
 
 static int rtw_sub_pwr_seq_parser(struct rtw_dev *rtwdev, u8 intf_mask,
-				  u8 cut_mask,
+				  u16 cut_mask,
 				  const struct rtw_pwr_seq_cmd *cmd)
 {
 	const struct rtw_pwr_seq_cmd *cur_cmd;
@@ -230,7 +230,7 @@ static int rtw_sub_pwr_seq_parser(struct rtw_dev *rtwdev, u8 intf_mask,
 int rtw_pwr_seq_parser(struct rtw_dev *rtwdev,
 		       const struct rtw_pwr_seq_cmd * const *cmd_seq)
 {
-	u8 cut_mask;
+	u16 cut_mask;
 	u8 intf_mask;
 	u8 cut;
 	u32 idx = 0;
